application/consumer/requirements.txt
aiofiles==0.4.0
aiokafka==0.4.3
aiopg==0.15.0
aioredis==1.2.0
async-timeout==3.0.1
cassandra-driver==3.16.0
hiredis==0.3.1
httptools==0.0.11
kafka-python==1.4.3
multidict==4.5.2
peewee==3.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
redis==3.0.1
sanic==0.8.3
six==1.12.0
SQLAlchemy==1.2.15
ujson==1.35
uvloop==0.11.3
websockets==5.0.1
